# zstd: faster next state update in BMI2 version of decode